+++ tools/cargo-miden/src/commands/mod.rs
@@ -1,7 +1,9 @@
 pub mod build;
 pub mod new_project;
+pub mod package_cache;
 pub mod test;
 
 pub use build::BuildCommand;
 pub use new_project::NewCommand;
+pub use package_cache::PackageCacheCommand;
 pub use test::TestCommand;